requirements.txt
streamlit
requests
beautifulsoup4
lxml
Pillow